requirements.txt
# TomatoCare - Required Python Packages
# Install with: pip install -r requirements.txt
# NOTE: Install PyTorch separately first (see below)

# ── PyTorch (install separately with CUDA support) ──
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
torch
torchvision
torchaudio

# ── TensorBoard ──
tensorboard>=2.14.0

# ── Image Processing ──
opencv-python>=4.8.0
Pillow>=10.0.0

# ── Data Augmentation ──
albumentations>=1.3.0

# ── Data Handling ──
numpy>=1.24.0
pandas>=2.0.0

# ── Visualization ──
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Machine Learning Utilities ──
scikit-learn>=1.3.0

# ── Jupyter Notebooks ──
jupyter>=1.0.0
ipykernel>=6.0.0

# ── Progress Bars ──
tqdm>=4.65.0
